requirements.txt
# RAG System with Gemini API - Dependencies

# Gemini API
google-generativeai>=0.8.0

# Vector Database
chromadb>=0.4.0

# Document Processing
PyPDF2>=3.0.0
python-docx>=0.8.0
markdown>=3.0.0

# Text Processing & NLP
nltk>=3.8.0
sentence-transformers>=2.2.0

# Web Interface
streamlit>=1.28.0
fastapi>=0.100.0
uvicorn>=0.20.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
requests>=2.28.0
aiohttp>=3.8.0

# Development & Testing
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0

# Optional: Local embeddings fallback
# transformers==4.35.0
# torch==2.1.0